+++ app/Services/Synchronization/SubGroupService.php
@@ -13,34 +13,36 @@ class SubGroupService
     {
         $this->repository = new SubGroupRepository();
         $this->syncRepository = new SyncRepository();
-        $this->syncEndpoint = env('API_URL').'pdv/carga-subgrupos?token=' . env('API_TOKEN');
+        $this->syncEndpoint = env('API_URL') . 'pdv/carga-subgrupos?token=' . env('API_TOKEN');
     }
 
     public function sync()
     {
         $get = Http::get($this->syncEndpoint)->json();
-        $response = $this->repository->all();
-        foreach ($response as $key => $r) {
-            $r->delete();
-        }
+        if ($get['code'] == '200') {
+            $response = $this->repository->all();
+            foreach ($response as $key => $r) {
+                $r->delete();
+            }
 
-        foreach ($get['data'] as $key => $data) {
-            $attributes = [
-                'idUsuario' => $data['id'],
-                'CodSistema' => $data['id'],
-                'codGrupo' => $data['grupo_id'],
-                'SubGrupo' => $data['nome'],
-            ];
+            foreach ($get['data'] as $key => $data) {
+                $attributes = [
+                    'idUsuario' => $data['id'],
+                    'CodSistema' => $data['id'],
+                    'codGrupo' => $data['grupo_id'],
+                    'SubGrupo' => $data['nome'],
+                ];
 
-            $this->repository->create($attributes);
-            Sync::updateOrCreate(
-                [
-                    'Tabela' => 'Subgrupo',
-                ],
-                [
-                    'DataSincronia' => date('Y-m-d H:i:s'),
-                ]
-            );
+                $this->repository->create($attributes);
+                Sync::updateOrCreate(
+                    [
+                        'Tabela' => 'Subgrupo',
+                    ],
+                    [
+                        'DataSincronia' => date('Y-m-d H:i:s'),
+                    ]
+                );
+            }
         }
     }
 }
