+++ app/Services/Synchronization/UserService.php
@@ -13,42 +13,44 @@ class UserService
     {
         $this->repository = new UserRepository();
         $this->syncRepository = new SyncRepository();
-        $this->syncEndpoint = env('API_URL').'pdv/carga-usuarios?token=' . env('API_TOKEN');
+        $this->syncEndpoint = env('API_URL') . 'pdv/carga-usuarios?token=' . env('API_TOKEN');
     }
 
     public function sync()
     {
         $get = Http::get($this->syncEndpoint)->json();
-        $response = $this->repository->all();
-        foreach ($response as $key => $r) {
-            $r->delete();
-        }
+        if ($get['code'] == '200') {
+            $response = $this->repository->all();
+            foreach ($response as $key => $r) {
+                $r->delete();
+            }
 
-        foreach ($get['data'] as $key => $data) {
-            $attributes = [
-                'idUsuario' => $data['id'],
-                'CodSistema' => $data['id'],
-                'Usuario' => $data['name'],
-                'Senha' => 123,
-                'flgUsuario' => 1,
-                'flgComissionado',
-                'Nome',
-                'flgAdmin',
-                'flgSupervisor',
-                'SenhaSupervisor',
-                'Codigo',
-                'Contato',
-            ];
+            foreach ($get['data'] as $key => $data) {
+                $attributes = [
+                    'idUsuario' => $data['id'],
+                    'CodSistema' => $data['id'],
+                    'Usuario' => $data['name'],
+                    'Senha' => 123,
+                    'flgUsuario' => 1,
+                    'flgComissionado',
+                    'Nome',
+                    'flgAdmin',
+                    'flgSupervisor',
+                    'SenhaSupervisor',
+                    'Codigo',
+                    'Contato',
+                ];
 
-            $this->repository->create($attributes);
-            Sync::updateOrCreate(
-                [
-                    'Tabela' => 'Usuario',
-                ],
-                [
-                    'DataSincronia' => date('Y-m-d H:i:s'),
-                ]
-            );
+                $this->repository->create($attributes);
+                Sync::updateOrCreate(
+                    [
+                        'Tabela' => 'Usuario',
+                    ],
+                    [
+                        'DataSincronia' => date('Y-m-d H:i:s'),
+                    ]
+                );
+            }
         }
     }
 }
